crm_backend/requirements-dev.txt
# Django Development Requirements

# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-environ==0.21.0

# Database
psycopg2-binary==2.9.9

# Task Queue
celery==5.3.4
redis==5.0.1
python-dateutil==2.8.2

# Google Sheets Integration
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.107.0

# File Processing
openpyxl==3.11.0

# Environment Variables
python-decouple==3.8

# Development Tools
black==23.12.1
flake8==6.1.0
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
